% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_borough_data2.R
\name{aggregate_borough_data2}
\alias{aggregate_borough_data2}
\title{Create borough data from ward model outputs}
\usage{
aggregate_borough_data2(model_output, constraint_path, first_proj_yr, n_cores)
}
\arguments{
\item{model_output}{List the output from the arrange_outputs function}

\item{constraint_path}{String the location of inflow and outflow borough data}

\item{first_proj_yr}{Numeric}

\item{n_cores}{Numeric. Number of cores}
}
\value{
A list of projection outputs by component for boroughs
}
\description{
For most components this is a simple sum function. For migration flows
scaling is applied to input flows so that the net matches the output from
the ward model but the gross flows are not just summed but are closer to
something reasonable. This is an attempt to account for intra-borough moves
in the ward data.
}
