% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_small_area_outputs.R
\name{arrange_small_area_outputs}
\alias{arrange_small_area_outputs}
\title{Arrange ward-level outputs}
\usage{
arrange_small_area_outputs(
  projection,
  population,
  births,
  deaths,
  in_migration,
  out_migration,
  fertility_rates,
  mortality_rates,
  in_mig_flows,
  out_mig_rates,
  dwelling_trajectory,
  dwelling_stock,
  first_proj_yr,
  last_proj_yr,
  config_list,
  model,
  n_cores
)
}
\arguments{
\item{projection}{A list output from housing_led_core}

\item{population}{A dataframe of the population backseries}

\item{births}{A dataframe of the births backseries}

\item{deaths}{A dataframe of the deaths backseries}

\item{in_migration}{A dataframe of the in migration backseries}

\item{out_migration}{A dataframe of the out migration backseries}

\item{fertility_rates}{A dataframe of fertility rates}

\item{mortality_rates}{A dataframe of mortality rates}

\item{in_mig_flows}{A dataframe of input in migration flows}

\item{out_mig_rates}{A dataframe of input out migration rates}

\item{dwelling_trajectory}{A dataframe}

\item{dwelling_stock}{A dataframe}

\item{first_proj_yr}{Numeric. The first projection year}

\item{last_proj_yr}{Numeric. The last projection}

\item{config_list}{List. The model config list}

\item{model}{String. Either 'trend' or 'housing-led'}

\item{n_cores}{Numeric. Number of cores}
}
\value{
A list of projection outputs by component
}
\description{
Take the projection output list which is organised by year and rearrange it
so that its organised by component. Add on the backseries.
}
