% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpo_template_to_rds.R
\name{bpo_template_to_rds}
\alias{bpo_template_to_rds}
\title{Process a standard bpo template csv file into 2 rds files}
\usage{
bpo_template_to_rds(csv_name, bpo_dir, trajectory_range, wards)
}
\arguments{
\item{csv_name}{String. The name of the dwelling trajectory csv saved in the
\code{bpo_dir} folder. With or without the "csv" suffix.}

\item{bpo_dir}{String. The folder containing the dwelling trajectory csv.}

\item{trajectory_range}{Numeric. The years covered by the input trajectory.}

\item{wards}{String indicating the ward geography in the projection: either WD13 or WD22}
}
\description{
Take a standard ward-level csv bpo template and create 2
rds files - 1 ward and 1 borough - where additional data
not contained in the template is taken from the SHLAA
}
