% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regross.R
\name{.optimise_gross_flows}
\alias{.optimise_gross_flows}
\title{Optimise gross flows using Poisson distributions}
\usage{
.optimise_gross_flows(base_in, base_out, target_net)
}
\arguments{
\item{base_in}{Dataframe}

\item{base_out}{Dataframe}

\item{target_net}{Dataframe}
}
\description{
Optimise gross flows using Poisson distributions
}
