% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regross.R
\name{.optimise_gross_flows_2}
\alias{.optimise_gross_flows_2}
\title{Optimise gross flows using Poisson distributions. Faster implementation.}
\usage{
.optimise_gross_flows_2(base_in, base_out, target_net, jump_scale = 4)
}
\arguments{
\item{base_in}{Dataframe}

\item{base_out}{Dataframe}

\item{target_net}{Dataframe}

\item{jump_scale}{Numeric. Lower values of jump scale should be faster, but
potentially less reliable. Always use a value greater than 1. Default = 4.}
}
\description{
Optimise gross flows using Poisson distributions. Faster implementation.
}
