% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regross.R
\name{.regross_parallel}
\alias{.regross_parallel}
\title{Regrosser in paralell}
\usage{
.regross_parallel(mig_target, fun = 2, n_cores = 8)
}
\arguments{
\item{mig_target}{Dataframe}

\item{fun}{Numeric. 1 or 2. Version of the regrosser function to use. Default 2.}

\item{n_cores}{Numeric. Number of cores. Default 8}
}
\description{
Regrosser in paralell
}
