% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regross.R
\name{.regross_serial}
\alias{.regross_serial}
\title{Regrosser in serial}
\usage{
.regross_serial(mig_target, fun = 2)
}
\arguments{
\item{mig_target}{Dataframe}

\item{fun}{Numeric. 1 or 2. Version of the regrosser function to use. Default 2.}
}
\description{
Regrosser in serial
}
