% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.unnest_dt}
\alias{.unnest_dt}
\title{DT implementation of tidyr::unnest()
https://www.r-bloggers.com/2019/10/much-faster-unnesting-with-data-table/}
\usage{
.unnest_dt(tbl, col)
}
\arguments{
\item{tbl}{Dataframe}

\item{col}{String}
}
\description{
DT implementation of tidyr::unnest()
https://www.r-bloggers.com/2019/10/much-faster-unnesting-with-data-table/
}
