% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmodel_hl_projection.R,
%   R/housing-led_run_proj.R
\name{flexmodel_hl_projection}
\alias{flexmodel_hl_projection}
\title{Run the small area housing-led model}
\usage{
flexmodel_hl_projection(config_list, n_cores = NULL)

flexmodel_hl_projection(config_list, n_cores = NULL)
}
\arguments{
\item{config_list}{A List. A housing-led model configuration list.}

\item{n_cores}{An integer describing the number of CPU cores to use}
}
\description{
Given a set of input parameters run a housing-led for small areas

Given a set of input parameters run a housing-led for small areas
}
