% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_run_proj.R
\name{flexmodel_trend_projection}
\alias{flexmodel_trend_projection}
\title{Run the small area trend model}
\usage{
flexmodel_trend_projection(config_list)
}
\arguments{
\item{config_list}{A List. A housing-led model configuration list.}
}
\description{
Given a set of input parameters run a cohort component model for small areas
}
