% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_led_core.R
\name{housing_led_core}
\alias{housing_led_core}
\title{Project from a starting population using a housing-led methodology to
produce an output population and components}
\usage{
housing_led_core(
  start_population,
  trend_projection,
  communal_establishment_population,
  household_rep_rates,
  households,
  projection_year,
  ahs_mix,
  hhr_ahs_uplift,
  constraint_list,
  n_cores,
  lookup
)
}
\arguments{
\item{start_population}{A dataframe. The starting population}

\item{trend_projection}{A list. The output from the small area trend model for
the year \code{projection_year}}

\item{communal_establishment_population}{A dataframe. The communal establishment
population}

\item{household_rep_rates}{A dataframe. Pre-calculated HHRs for model areas}

\item{households}{A dataframe. An input total household figure for model areas}

\item{projection_year}{Numeric. The year being projected}

\item{ahs_mix}{Numeric. A number between 0 and 1 defining the mix of AHS
from the HHR household model and the implied trend AHS}

\item{hhr_ahs_uplift}{Numeric or NULL. The difference between the input AHS
and the trend AHS in the first year of the projection.}

\item{constraint_list}{A list of constraints information and data for use
in the model loop.}

\item{n_cores}{Numeric. Number of cores}

\item{lookup}{Dataframe. gss_code_ward to ward_name (cols must be named like this)}
}
\value{
A list of projected components
}
\description{
Project from a starting population using a housing-led methodology to
produce an output population and components
}
