% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_small_area_projection.R
\name{output_small_area_projection}
\alias{output_small_area_projection}
\title{Write model outputs to RDS and CSV}
\usage{
output_small_area_projection(
  projection,
  output_dir = config_list$output_dir,
  model
)
}
\arguments{
\item{projection}{A list. The projection as processed through the arrange function}

\item{output_dir}{String. The path to the output folder where files will be written}

\item{model}{A string. Either 'trend' or 'housing-led' indicating what data is
being passed to the function. Used to select which list elements to output as CSV.}
}
\description{
Outputs from the small area housing-led model, passed through the \code{arrange}
function are written out a RDS and CSV files
}
