% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regross_parallel.R
\name{regross_parallel}
\alias{regross_parallel}
\title{Regrosser}
\usage{
regross_parallel(
  base_in,
  base_out,
  target_net,
  col_inflow = "inflow",
  col_outflow = "outflow",
  col_target = "net_target",
  n_cores = 8
)
}
\arguments{
\item{base_in}{Dataframe}

\item{base_out}{Dataframe}

\item{target_net}{Dataframe}

\item{col_inflow}{String}

\item{col_outflow}{String}

\item{col_target}{String}

\item{n_cores}{Numeric. Number of cores}
}
\description{
Regrosser
}
