% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bpo_projection.R
\name{run_bpo_projection}
\alias{run_bpo_projection}
\title{Run the borough and ward housing-led models using a BPO dwelling trajectory}
\usage{
run_bpo_projection(
  bpo_name,
  wards,
  variant,
  trajectory_range = 2012:2041,
  projection_range = 2021:2041,
  bpo_dir = "Q:/Teams/D&PA/Demography/Projections/bpo_2020_based/",
  csv_name = bpo_name
)
}
\arguments{
\item{bpo_name}{String. The name of the dwelling trajectory csv saved in the
folder \code{bpo_dir} folder. The output projection will also have this name.}

\item{wards}{String. Either WD13 or WD22}

\item{variant}{String. Either lower or upper}

\item{trajectory_range}{Numeric or NULL. The years covered by the input trajectory.
Default 2020:2041.}

\item{projection_range}{Numeric. The years to project. \code{Default 2020:2050}.}

\item{bpo_dir}{String. The folder containing the dwelling trajectory csv}

\item{csv_name}{String. The name of the dwelling trajectory csv saved in the}
}
\description{
A wrapper for run_borough_and_ward_projection. Takes as input a csv in the
of the GLA housing trajectory template for any borough and converts it into
an rds in tidy format. Run the borough housing-led model and the ward small
area model using a set of standard model parameters.
}
