% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_run_proj.R
\name{run_small_area_trend_model}
\alias{run_small_area_trend_model}
\title{Run the small area trend model}
\usage{
run_small_area_trend_model(config_list)
}
\arguments{
\item{config_list}{A List. A housing-led model configuration list.}
}
\description{
Given a set of input parameters run a cohort component model for small areas
}
